# chore: speed up build.js