```
hyperfine --parameter-list branch main,chore/build-optimization --setup "git switch {branch} && npm i && npm run build" "npm run build" --runs 10
Benchmark 1: npm run build (branch = main)
  Time (mean ± σ):      7.249 s ±  0.081 s    [User: 4.068 s, System: 0.589 s]
  Range (min … max):    7.144 s …  7.394 s    10 runs
 
Benchmark 2: npm run build (branch = chore/build-optimization)
  Time (mean ± σ):      4.998 s ±  0.066 s    [User: 3.907 s, System: 0.412 s]
  Range (min … max):    4.920 s …  5.115 s    10 runs
 
Summary
  npm run build (branch = chore/build-optimization) ran
    1.45 ± 0.03 times faster than npm run build (branch = main) 
```